[Order] Move CompositeOrderProcessor and OrderProcessorInterface to Order component
