+++ src/Sylius/Component/Core/OrderProcessing/OrderAdjustmentsClearer.php
@@ -12,7 +12,8 @@
 namespace Sylius\Component\Core\OrderProcessing;
 
 use Sylius\Component\Core\Model\AdjustmentInterface;
-use Sylius\Component\Core\Model\OrderInterface;
+use Sylius\Component\Order\Model\OrderInterface;
+use Sylius\Component\Order\Processor\OrderProcessorInterface;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
