+++ src/Sylius/Component/Core/OrderProcessing/OrderPricesRecalculator.php
@@ -11,7 +11,8 @@
 
 namespace Sylius\Component\Core\OrderProcessing;
 
-use Sylius\Component\Core\Model\OrderInterface;
+use Sylius\Component\Order\Model\OrderInterface;
+use Sylius\Component\Order\Processor\OrderProcessorInterface;
 use Sylius\Component\Pricing\Calculator\DelegatingCalculatorInterface;
 
 /**
