+++ src/Sylius/Component/Core/spec/OrderProcessing/OrderAdjustmentsClearerSpec.php
@@ -15,7 +15,7 @@ use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\AdjustmentInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\OrderProcessing\OrderAdjustmentsClearer;
-use Sylius\Component\Core\OrderProcessing\OrderProcessorInterface;
+use Sylius\Component\Order\Processor\OrderProcessorInterface;
 
 /**
  * @mixin OrderAdjustmentsClearer
