+++ src/Sylius/Component/Core/spec/OrderProcessing/OrderExchangeRateProcessorSpec.php
@@ -15,7 +15,7 @@ use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\OrderProcessing\OrderExchangeRateProcessor;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Sylius\Component\Core\OrderProcessing\OrderProcessorInterface;
+use Sylius\Component\Order\Processor\OrderProcessorInterface;
 use Sylius\Component\Core\Updater\OrderUpdaterInterface;
 
 /**
